feat: rename disabilityStatus to hasDisability
